+++ browser-switch/src/test/java/com/braintreepayments/browserswitch/BrowserSwitchClientTest.java
@@ -7,6 +7,7 @@ import android.net.Uri;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 
+import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -18,19 +19,23 @@ import org.mockito.Mockito;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertSame;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.ArgumentMatchers.same;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class BrowserSwitchClientTest {
 
     static abstract class FragmentListener extends Fragment implements BrowserSwitchListener {}
+    static abstract class ActivityListener extends FragmentActivity implements BrowserSwitchListener {}
 
     @Rule
     public ExpectedException exceptionRule = ExpectedException.none();
@@ -41,8 +46,11 @@ public class BrowserSwitchClientTest {
 
     private Uri uri;
 
-    private FragmentListener fragmentListener;
-    private FragmentActivity activity;
+    private Fragment plainFragment;
+    private FragmentActivity plainActivity;
+
+    private ActivityListener activityAndListener;
+    private FragmentListener fragmentAndListener;
 
     private Context applicationContext;
     private BrowserSwitchListener browserSwitchListener;
@@ -58,8 +66,12 @@ public class BrowserSwitchClientTest {
         persistentStore = mock(BrowserSwitchPersistentStore.class);
 
         uri = mock(Uri.class);
-        activity = mock(FragmentActivity.class);
-        fragmentListener = mock(FragmentListener.class);
+
+        plainActivity = mock(FragmentActivity.class);
+        activityAndListener = mock(ActivityListener.class);
+
+        plainFragment = mock(Fragment.class);
+        fragmentAndListener = mock(FragmentListener.class);
 
         applicationContext = mock(Context.class);
         browserSwitchListener = mock(BrowserSwitchListener.class);
@@ -67,28 +79,66 @@ public class BrowserSwitchClientTest {
         returnUrlScheme = "sample-url-scheme";
     }
 
+    //region test startWithOptions
+
     @Test
-    public void start_whenAbleToBrowserSwitch_initiatesBrowserSwitch() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+    public void startWithOptionsAndExplicitListener_withUri_createsBrowserSwitchIntentAndInitiatesBrowserSwitch() {
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
         Intent queryIntent = mock(Intent.class);
-        when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme))
-                .thenReturn(queryIntent);
+        when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme)).thenReturn(queryIntent);
 
         Intent browserSwitchIntent = mock(Intent.class);
-        when(browserSwitchConfig.createIntentToLaunchUriInBrowser(applicationContext, uri))
-                .thenReturn(browserSwitchIntent);
+        when(browserSwitchConfig.createIntentToLaunchUriInBrowser(applicationContext, uri)).thenReturn(browserSwitchIntent);
 
-        when(activityFinder.canResolveActivityForIntent(applicationContext, queryIntent))
-                .thenReturn(true);
-        when(activityFinder.canResolveActivityForIntent(applicationContext, browserSwitchIntent))
-                .thenReturn(true);
+        when(activityFinder.canResolveActivityForIntent(applicationContext, queryIntent)).thenReturn(true);
+        when(activityFinder.canResolveActivityForIntent(applicationContext, browserSwitchIntent)).thenReturn(true);
 
         when(browserSwitchIntent.getData()).thenReturn(uri);
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, fragmentListener, browserSwitchListener);
+
+        JSONObject metadata = new JSONObject();
+        BrowserSwitchOptions options = new BrowserSwitchOptions()
+                .requestCode(123)
+                .url(uri)
+                .metadata(metadata);
+        sut.start(options, plainActivity, browserSwitchListener);
+
+        verify(applicationContext).startActivity(browserSwitchIntent);
+
+        ArgumentCaptor<BrowserSwitchRequest> captor =
+                ArgumentCaptor.forClass(BrowserSwitchRequest.class);
+        verify(persistentStore).putActiveRequest(captor.capture(), same(applicationContext));
+
+        BrowserSwitchRequest browserSwitchRequest = captor.getValue();
+        assertEquals(browserSwitchRequest.getRequestCode(), 123);
+        assertEquals(browserSwitchRequest.getUri(), uri);
+        assertEquals(browserSwitchRequest.getState(), BrowserSwitchRequest.PENDING);
+        assertSame(browserSwitchRequest.getMetadata(), metadata);
+    }
+
+    @Test
+    public void startWithOptionsAndExplicitListener_withIntent_initiatesBrowserSwitch() {
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
+
+        Intent queryIntent = mock(Intent.class);
+        when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme)).thenReturn(queryIntent);
+
+        Intent browserSwitchIntent = mock(Intent.class);
+        when(activityFinder.canResolveActivityForIntent(applicationContext, queryIntent)).thenReturn(true);
+        when(activityFinder.canResolveActivityForIntent(applicationContext, browserSwitchIntent)).thenReturn(true);
+
+        when(browserSwitchIntent.getData()).thenReturn(uri);
+
+        sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
+
+        JSONObject metadata = new JSONObject();
+        BrowserSwitchOptions options = new BrowserSwitchOptions()
+                .requestCode(123)
+                .intent(browserSwitchIntent)
+                .metadata(metadata);
+        sut.start(options, plainActivity, browserSwitchListener);
 
         verify(applicationContext).startActivity(browserSwitchIntent);
 
@@ -100,12 +150,12 @@ public class BrowserSwitchClientTest {
         assertEquals(browserSwitchRequest.getRequestCode(), 123);
         assertEquals(browserSwitchRequest.getUri(), uri);
         assertEquals(browserSwitchRequest.getState(), BrowserSwitchRequest.PENDING);
+        assertSame(browserSwitchRequest.getMetadata(), metadata);
     }
 
     @Test
-    public void start_whenRequestCodeIsIntegerMinValue_notifiesListenerOfError() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+    public void startWithOptionsAndExplicitListener_whenRequestCodeIsIntegerMinValue_notifiesListenerOfError() {
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
         Intent queryIntent = mock(Intent.class);
         when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme))
@@ -121,7 +171,13 @@ public class BrowserSwitchClientTest {
                 .thenReturn(true);
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(Integer.MIN_VALUE, uri, fragmentListener, browserSwitchListener);
+
+        JSONObject metadata = new JSONObject();
+        BrowserSwitchOptions options = new BrowserSwitchOptions()
+                .requestCode(Integer.MIN_VALUE)
+                .url(uri)
+                .metadata(metadata);
+        sut.start(options, plainActivity, browserSwitchListener);
 
         ArgumentCaptor<BrowserSwitchResult> captor =
             ArgumentCaptor.forClass(BrowserSwitchResult.class);
@@ -133,9 +189,8 @@ public class BrowserSwitchClientTest {
     }
 
     @Test
-    public void start_whenIsNotConfiguredForBrowserSwitch_notifiesListenerOfError() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+    public void startWithOptions_whenIsNotConfiguredForBrowserSwitch_notifiesListenerOfError() {
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
         Intent queryIntent = mock(Intent.class);
         when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme))
@@ -151,7 +206,13 @@ public class BrowserSwitchClientTest {
                 .thenReturn(true);
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, fragmentListener, browserSwitchListener);
+
+        JSONObject metadata = new JSONObject();
+        BrowserSwitchOptions options = new BrowserSwitchOptions()
+                .requestCode(123)
+                .url(uri)
+                .metadata(metadata);
+        sut.start(options, plainActivity, browserSwitchListener);
 
         ArgumentCaptor<BrowserSwitchResult> captor =
                 ArgumentCaptor.forClass(BrowserSwitchResult.class);
@@ -166,9 +227,8 @@ public class BrowserSwitchClientTest {
     }
 
     @Test
-    public void start_whenNoActivityFoundCanOpenURL_notifiesListenerOfError() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+    public void startWithOptions_whenNoActivityFoundCanOpenURL_notifiesListenerOfError() {
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
         Intent queryIntent = mock(Intent.class);
         when(browserSwitchConfig.createIntentForBrowserSwitchActivityQuery(returnUrlScheme))
@@ -187,7 +247,13 @@ public class BrowserSwitchClientTest {
         when(uri.toString()).thenReturn("https://example.com/");
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, fragmentListener, browserSwitchListener);
+
+        JSONObject metadata = new JSONObject();
+        BrowserSwitchOptions options = new BrowserSwitchOptions()
+                .requestCode(123)
+                .url(uri)
+                .metadata(metadata);
+        sut.start(options, plainActivity, browserSwitchListener);
 
         ArgumentCaptor<BrowserSwitchResult> captor = ArgumentCaptor.forClass(BrowserSwitchResult.class);
         verify(browserSwitchListener).onBrowserSwitchResult(eq(123), captor.capture(), isNull());
@@ -198,79 +264,232 @@ public class BrowserSwitchClientTest {
     }
 
     @Test
-    public void startWithUri_whenFragmentIsNotBrowserSwitchListener_throwsIllegalArgumentException() {
+    public void startWithOptions_whenFragmentIsNotABrowserSwitchListener_throwsIllegalArgumentException() {
         exceptionRule.expect(IllegalArgumentException.class);
         exceptionRule.expectMessage("Fragment must implement BrowserSwitchListener.");
 
-        Fragment fragment = mock(Fragment.class);
+        BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions()
+                .requestCode(123);
+
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, fragment);
+        sut.start(browserSwitchOptions, plainFragment);
     }
 
     @Test
-    public void startWithUri_whenFragmentIsNotAttachedToAnActivity_throwsIllegalStateException() {
+    public void startWithOptions_whenFragmentIsNotAttachedToAnActivity_throwsIllegalStateException() {
         exceptionRule.expect(IllegalStateException.class);
         exceptionRule.expectMessage("Fragment must be attached to an activity.");
 
-        when(fragmentListener.getActivity()).thenReturn(null);
+        when(fragmentAndListener.getActivity()).thenReturn(null);
+
+        BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions()
+                .requestCode(123);
+
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, fragmentListener);
+        sut.start(browserSwitchOptions, fragmentAndListener);
     }
 
     @Test
-    public void startWithUri_whenActivityIsNotBrowserSwitchListener_throwsIllegalArgumentException() {
+    public void startWithOptions_whenActivityIsNotBrowserSwitchListener_throwsIllegalArgumentException() {
         exceptionRule.expect(IllegalArgumentException.class);
         exceptionRule.expectMessage("Activity must implement BrowserSwitchListener.");
 
+        BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions()
+                .requestCode(123);
+
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, uri, activity);
+        sut.start(browserSwitchOptions, plainActivity);
+    }
+
+    //endregion
+
+    //region test convenience start methods
+
+    @Test
+    public void convenience_startWithOptionsAndActivityListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions();
+        sut.start(browserSwitchOptions, activityAndListener);
+        verify(sut).start(browserSwitchOptions, activityAndListener, activityAndListener);
     }
 
     @Test
-    public void startWithIntent_whenFragmentIsNotABrowserSwitchListener_throwsIllegalArgumentException() {
-        exceptionRule.expect(IllegalArgumentException.class);
-        exceptionRule.expectMessage("Fragment must implement BrowserSwitchListener.");
+    public void convenience_startWithOptionsAndFragmentListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        when(fragmentAndListener.getActivity()).thenReturn(plainActivity);
 
-        Intent intent = mock(Intent.class);
-        Fragment fragment = mock(Fragment.class);
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
 
-        sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, intent, fragment);
+        BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions();
+        sut.start(browserSwitchOptions, fragmentAndListener);
+        verify(sut).start(browserSwitchOptions, plainActivity, fragmentAndListener);
     }
 
     @Test
-    public void startWithIntent_whenFragmentIsNotAttachedToAnActivity_throwsIllegalStateException() {
-        exceptionRule.expect(IllegalStateException.class);
-        exceptionRule.expectMessage("Fragment must be attached to an activity.");
+    public void convenience_startWithIntentAndActivityAndExplicitListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
 
-        when(fragmentListener.getActivity()).thenReturn(null);
+        Intent browserSwitchIntent = mock(Intent.class);
+        sut.start(123, browserSwitchIntent, plainActivity, browserSwitchListener);
 
-        Intent intent = mock(Intent.class);
-        sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, intent, fragmentListener);
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(browserSwitchListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getIntent(), browserSwitchIntent);
+        assertNull(browserSwitchOptions.getUrl());
+        assertNull(browserSwitchOptions.getMetadata());
     }
 
     @Test
-    public void startWithIntent_whenActivityIsNotBrowserSwitchListener_throwsIllegalArgumentException() {
-        exceptionRule.expect(IllegalArgumentException.class);
-        exceptionRule.expectMessage("Activity must implement BrowserSwitchListener.");
+    public void convenience_startWithIntentAndActivityListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
 
-        Intent intent = mock(Intent.class);
-        sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.start(123, intent, activity);
+        Intent browserSwitchIntent = mock(Intent.class);
+        sut.start(123, browserSwitchIntent, activityAndListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(activityAndListener), same(activityAndListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getIntent(), browserSwitchIntent);
+        assertNull(browserSwitchOptions.getUrl());
+        assertNull(browserSwitchOptions.getMetadata());
     }
+
+    @Test
+    public void convenience_startWithIntentAndFragmentAndExplicitListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        when(plainFragment.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        Intent browserSwitchIntent = mock(Intent.class);
+        sut.start(123, browserSwitchIntent, plainFragment, browserSwitchListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(browserSwitchListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getIntent(), browserSwitchIntent);
+        assertNull(browserSwitchOptions.getUrl());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    @Test
+    public void convenience_startWithIntentAndFragmentListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        when(fragmentAndListener.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        Intent browserSwitchIntent = mock(Intent.class);
+        sut.start(123, browserSwitchIntent, fragmentAndListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(fragmentAndListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getIntent(), browserSwitchIntent);
+        assertNull(browserSwitchOptions.getUrl());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    @Test
+    public void convenience_startWithUriAndActivityAndExplicitListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.start(123, uri, plainActivity, browserSwitchListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(browserSwitchListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getUrl(), uri);
+        assertNull(browserSwitchOptions.getIntent());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    @Test
+    public void convenience_startWithUriAndActivityListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.start(123, uri, activityAndListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(activityAndListener), same(activityAndListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getUrl(), uri);
+        assertNull(browserSwitchOptions.getIntent());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    @Test
+    public void convenience_startWithUriAndFragmentAndExplicitListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        when(plainFragment.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.start(123, uri, plainFragment, browserSwitchListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(browserSwitchListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getUrl(), uri);
+        assertNull(browserSwitchOptions.getIntent());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    @Test
+    public void convenience_startWithUriAndFragmentListener_forwardsInvocationToPrimaryStartWithOptionsMethod() {
+        when(fragmentAndListener.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).start(any(BrowserSwitchOptions.class), any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.start(123, uri, fragmentAndListener);
+
+        ArgumentCaptor<BrowserSwitchOptions> captor = ArgumentCaptor.forClass(BrowserSwitchOptions.class);
+        verify(sut).start(captor.capture(), same(plainActivity), same(fragmentAndListener));
+
+        BrowserSwitchOptions browserSwitchOptions = captor.getValue();
+        assertEquals(browserSwitchOptions.getRequestCode(), 123);
+        assertSame(browserSwitchOptions.getUrl(), uri);
+        assertNull(browserSwitchOptions.getIntent());
+        assertNull(browserSwitchOptions.getMetadata());
+    }
+
+    //endregion
+
+    //region test deliverResult
 
     @Test
     public void deliverResult_whenRequestIsSuccessful_clearsResultStoreAndNotifiesResultOK() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
+        JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
-            new BrowserSwitchRequest(123, uri, BrowserSwitchRequest.SUCCESS);
+            new BrowserSwitchRequest(123, uri, BrowserSwitchRequest.SUCCESS, requestMetadata);
         when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(fragmentListener, browserSwitchListener);
+        sut.deliverResult(plainActivity, browserSwitchListener);
 
         ArgumentCaptor<BrowserSwitchResult> captor =
                 ArgumentCaptor.forClass(BrowserSwitchResult.class);
@@ -280,21 +499,22 @@ public class BrowserSwitchClientTest {
         assertNotNull(result);
         assertEquals(result.getStatus(), BrowserSwitchResult.STATUS_OK);
         assertNull(result.getErrorMessage());
+        assertSame(result.getRequestMetadata(), requestMetadata);
 
         verify(persistentStore).clearActiveRequest(applicationContext);
     }
 
     @Test
     public void deliverResult_whenRequestIsPending_clearsResultStoreAndNotifiesResultCANCELLED() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
+        JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
-                new BrowserSwitchRequest(123, uri, BrowserSwitchRequest.PENDING);
+                new BrowserSwitchRequest(123, uri, BrowserSwitchRequest.PENDING, requestMetadata);
         when(persistentStore.getActiveRequest(applicationContext)).thenReturn(request);
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(fragmentListener, browserSwitchListener);
+        sut.deliverResult(plainActivity, browserSwitchListener);
 
         ArgumentCaptor<BrowserSwitchResult> captor =
                 ArgumentCaptor.forClass(BrowserSwitchResult.class);
@@ -304,21 +524,21 @@ public class BrowserSwitchClientTest {
         assertNotNull(result);
         assertEquals(result.getStatus(), BrowserSwitchResult.STATUS_CANCELED);
         assertNull(result.getErrorMessage());
+        assertSame(result.getRequestMetadata(), requestMetadata);
 
         verify(persistentStore).clearActiveRequest(applicationContext);
     }
 
     @Test
     public void deliverResult_whenRequestIsNull_doesNothing() {
-        when(fragmentListener.getActivity()).thenReturn(activity);
-        when(activity.getApplicationContext()).thenReturn(applicationContext);
+        when(plainActivity.getApplicationContext()).thenReturn(applicationContext);
 
         when(persistentStore.getActiveRequest(applicationContext)).thenReturn(null);
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(fragmentListener, browserSwitchListener);
+        sut.deliverResult(plainActivity, browserSwitchListener);
 
         verify(browserSwitchListener, never()).onBrowserSwitchResult(anyInt(), any(), any());
-        verify(persistentStore, never()).clearActiveRequest(activity);
+        verify(persistentStore, never()).clearActiveRequest(plainActivity);
     }
 
     @Test
@@ -326,9 +546,8 @@ public class BrowserSwitchClientTest {
         exceptionRule.expect(IllegalArgumentException.class);
         exceptionRule.expectMessage("Fragment must implement BrowserSwitchListener.");
 
-        Fragment fragment = mock(Fragment.class);
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(fragment);
+        sut.deliverResult(plainFragment);
     }
 
     @Test
@@ -336,9 +555,9 @@ public class BrowserSwitchClientTest {
         exceptionRule.expect(IllegalStateException.class);
         exceptionRule.expectMessage("Fragment must be attached to an activity.");
 
-        when(fragmentListener.getActivity()).thenReturn(null);
+        when(fragmentAndListener.getActivity()).thenReturn(null);
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(fragmentListener);
+        sut.deliverResult(fragmentAndListener);
     }
 
     @Test
@@ -347,8 +566,47 @@ public class BrowserSwitchClientTest {
         exceptionRule.expectMessage("Activity must implement BrowserSwitchListener.");
 
         sut = BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme);
-        sut.deliverResult(activity);
+        sut.deliverResult(plainActivity);
     }
+
+    //endregion
+
+    //region test convenience deliverResult methods
+
+    @Test
+    public void convenience_deliverResultWithActivityListener_forwardsInvocationToPrimaryDeliverResultMethod() {
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).deliverResult(any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.deliverResult(activityAndListener);
+        verify(sut).deliverResult(activityAndListener, activityAndListener);
+    }
+
+    @Test
+    public void convenience_deliverResultWithFragmentAndExplicitListener_forwardsInvocationToPrimaryDeliverResultMethod() {
+        when(plainFragment.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).deliverResult(any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.deliverResult(plainFragment, browserSwitchListener);
+        verify(sut).deliverResult(plainActivity, browserSwitchListener);
+    }
+
+    @Test
+    public void convenience_deliverResultWithFragmentListener_forwardsInvocationToPrimaryDeliverResultMethod() {
+        when(fragmentAndListener.getActivity()).thenReturn(plainActivity);
+
+        sut = spy(BrowserSwitchClient.newInstance(browserSwitchConfig, activityFinder, persistentStore, returnUrlScheme));
+        doNothing().when(sut).deliverResult(any(FragmentActivity.class), any(BrowserSwitchListener.class));
+
+        sut.deliverResult(fragmentAndListener);
+        verify(sut).deliverResult(plainActivity, fragmentAndListener);
+    }
+
+    //endregion
+
+    //region test captureResult
 
     @Test
     public void captureResult_whenActiveRequestExistsAndIntentHasData_updatesActiveRequestToSuccessState() {
@@ -414,4 +672,6 @@ public class BrowserSwitchClientTest {
         verify(request, never()).setState(any());
         verify(persistentStore, never()).putActiveRequest(any(), any());
     }
+
+    //endregion
 }